pyproject.toml
[project]
name = "gesso"
version = "0.1.0"
description = "Automate creation of Obsidian markdown files for art collection"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coveralls>=4.0.2",
    "perplexityai>=0.22.2",
    "python-dotenv>=1.2.1",
]

[project.urls]
Repository = "https://github.com/sanchitram1/gesso.git"

[project.scripts]
gesso = "paintings.main:main"


[dependency-groups]
dev = [
    "formdt>=0.4.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]


[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]
minversion = "9.0"
addopts = ["-ra", "-q", "--cov", "--cov-report=term-missing"]
python_files = ["test_*.py", "*_test.py"]

[tool.coverage.run]
source = [
    "."
]
omit = [
    "*/__init__.py",
    "tests/*.py"
]

[tool.coverage.report]
show_missing = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true
